Refactor request service tests into focused suites
